chore: Remove code duplication
