+++ functional-domain/src/main/java/io/github/fherbreteau/functional/domain/command/impl/check/AbstractCheckUserCommand.java
@@ -1,22 +1,16 @@
 package io.github.fherbreteau.functional.domain.command.impl.check;
 
-import static java.lang.System.Logger.Level.DEBUG;
-
 import java.util.List;
 
-import io.github.fherbreteau.functional.domain.command.CheckCommand;
 import io.github.fherbreteau.functional.domain.command.Command;
 import io.github.fherbreteau.functional.domain.command.impl.error.UserErrorCommand;
 import io.github.fherbreteau.functional.domain.entities.Output;
-import io.github.fherbreteau.functional.domain.entities.User;
 import io.github.fherbreteau.functional.driven.repository.GroupRepository;
 import io.github.fherbreteau.functional.driven.repository.UserRepository;
 import io.github.fherbreteau.functional.driven.rules.UserChecker;
 import io.github.fherbreteau.functional.driven.rules.UserUpdater;
 
-public abstract class AbstractCheckUserCommand<T, C extends Command<Output<T>>> implements CheckCommand<T> {
-    protected final System.Logger logger = System.getLogger(getClass().getSimpleName());
-
+public abstract class AbstractCheckUserCommand<T, C extends Command<Output<T>>> extends AbstractCheckCommand<T, C, UserErrorCommand<T>> {
     protected final UserRepository userRepository;
     protected final GroupRepository groupRepository;
     protected final UserChecker userChecker;
@@ -29,21 +23,6 @@ public abstract class AbstractCheckUserCommand<T, C extends Command<Output<T>>> 
         this.userChecker = userChecker;
         this.userUpdater = userUpdater;
     }
-
-    @Override
-    public final Command<Output<T>> execute(User actor) {
-        List<String> reasons = checkAccess(actor);
-        if (!reasons.isEmpty()) {
-            logger.log(DEBUG, "Command will fail for {0} because of {1}", actor, reasons);
-            return createError(reasons);
-        }
-        logger.log(DEBUG, "Command will continue for {0}", actor);
-        return createSuccess();
-    }
-
-    protected abstract List<String> checkAccess(User actor);
-
-    protected abstract C createSuccess();
 
     protected UserErrorCommand<T> createError(List<String> reasons) {
         throw new UnsupportedOperationException("Unsupported Command always succeed");
